Adding two new types to our Procfile:
    - scheduler, and
    - worker

scheduler:

This is a celery instance that also runs celerybeat (the periodic task
scheduler). Since this instance runs celerybeat--it should NEVER BE SCALED TO
MORE THAN 1 WORKER!!!

If it's scaled to multiple workers, then tasks will be duplicated, causing
unnecessary load / issues.

worker:

This is a generic celery queue worker that can be scaled to infinity (and
beyond)! to process queues.