Adding django-compressor to the project.
This commit includes some basic configuration options. This tool is really
useful because it lets us use template contexts in our CSS / JS files, eg:

    # some css code
    body {
        background: url({{ STATIC_URL }}img/background.jpg);
    }

For local development, this is what it allows us to do. For production, it
allows us to compress and upload our assets to S3--thereby making downloads
really fast, and elegant in all of our development environments.